tabung.py

+++ tabung.py
@@ -1,6 +1,6 @@
 r = int(input("Masukkan Jari - jari : "))
 t = int(input("Masukkan Tinggi : "))
 
-volume = float(3.14 * r * r * t)
+volume = 3.14 * r * r * t
 
 print(f'Hasilnya adalah : {volume:.2f}')
